+++ HunterPie.Core/Native/IPC/Handlers/Internal/Initialiaze/Models/IPCInitializationHostType.cs
@@ -1,0 +1,10 @@
+﻿namespace HunterPie.Core.Native.IPC.Handlers.Internal.Initialiaze.Models;
+
+public enum IPCInitializationHostType
+{
+    Invalid = 0,
+    /// <summary>Monster Hunter World, Monster Hunter World: Iceborne.</summary>
+    MHWorld,
+    /// <summary>Monster Hunter Rise, Monster Hunter World: Sunbreak.</summary>
+    MHRise,
+}
